requirements.txt
praw==7.7.1
pandas==2.0.3
nltk==3.8.1
streamlit==1.24.0
plotly==5.15.0
python-dotenv==1.0.0 